refactor: merge FaultManagementListView into FaultManagementView
